travel-guide-backend/requirements.txt
boto3>=1.26.137
python-jose[cryptography]>=3.3.0
requests>=2.31.0
Pillow>=10.0.0 